+++ src/main/java/com/lr/salestax/goods/impl/MedicalProducts.java
@@ -21,7 +21,7 @@ public class MedicalProducts
 	 * @param noOfItems
 	 * @param isImported
 	 */
-	public MedicalProducts(final String goodsDescription, final double goodsCostPerItem, final double noOfItems, final boolean isImported) {
+	public MedicalProducts(final String goodsDescription, final double goodsCostPerItem, final int noOfItems, final boolean isImported) {
 		super(goodsDescription, goodsCostPerItem, noOfItems, isImported);
 	}
